+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 **Fixed**
 
 * Fix regression where autoplay is not working
+* Fix system paths not being honored in macOS
 
 ## [Sabaki v0.17.0][v0.17.0] (2016-08-10)